--- order-service/src/main/resources/bootstrap.yml
@@ -1,6 +1,0 @@
-spring:
-  cloud:
-    config:
-      uri: http://localhost:8888
-      name: message
-      profile: native
