+++ user-service/src/main/resources/application.yml
@@ -4,6 +4,20 @@ server:
 spring:
   application:
     name: user-service
+  profiles:
+    active: local
+  config:
+    import: "configserver:http://localhost:8888"
+  cloud:
+    config: # Ex) `http://localhost:8088/config/test`
+      name: config # config name
+      profile: local # config profile
+
+management:
+  endpoints:
+    web:
+      exposure:
+        include: refresh
 
 eureka:
   instance:
